Typescript (#2133)
* setup initial typescript + eslint config (#1972)

* setup initial typescript + eslint config

* declare type for returning react component

* add loading prop options

* add tsconfig and eslint config

* add no emit flag in tsconfig

* delete emitted file

* apply code suggestions

* chore: rename types


* chore: update lint yml node version to v14 (#2131)


* chore: update eslint config to run typescript-eslint only for ts & tsx files (#2132)


* script: add .ts|tsx extensions for eslint script

* script: add .jsx extensions for eslint script

* chore: update typescript version
