+++ package.json
@@ -47,6 +47,12 @@
         "webpack": "^5.39.0"
     },
     "devDependencies": {
+        "@types/node": "^16.6.1",
+        "@types/react": "^17.0.18",
+        "@types/react-dom": "^17.0.9",
+        "@types/styled-components": "^5.1.11",
+        "@typescript-eslint/eslint-plugin": "^4.29.0",
+        "@typescript-eslint/parser": "^4.29.0",
         "babel-eslint": "10.1.0",
         "babel-plugin-styled-components": "^1.12.0",
         "@storybook/addon-actions": "^6.3.4",
@@ -59,7 +65,7 @@
         "eslint-loader": "^4.0.2",
         "eslint-plugin-import": "^2.23.4",
         "eslint-plugin-react": "^7.24.0",
-        "eslint-webpack-plugin": "^2.5.4",
+        "eslint-webpack-plugin": "^3.0.1",
         "gatsby-plugin-eslint": "^3.0.0",
         "gatsby-plugin-styled-components": "^4.7.1",
         "gh-pages": "^3.2.1",
@@ -74,7 +80,8 @@
         "stylelint-config-standard": "^22.0.0",
         "stylelint-config-styled-components": "^0.1.1",
         "stylelint-processor-styled-components": "^1.10.0",
-        "stylelint-webpack-plugin": "^2.2.0"
+        "stylelint-webpack-plugin": "^2.2.0",
+        "typescript": "^4.4.4"
     },
     "keywords": [
         "deriv",
@@ -86,7 +93,7 @@
         "develop": "gatsby develop -o",
         "format": "prettier --write '**/*.js' '*.js'",
         "stylelint": "stylelint 'src/**/*.js' --formatter verbose",
-        "eslint": "DEBUG=eslint:cli-engine eslint {'src/**/*.js','./*js'} --fix",
+        "eslint": "DEBUG=eslint:cli-engine eslint {'src/**/*.{js,jsx,ts,tsx}','./*.js'} --fix",
         "start": "npm run develop",
         "start-80": "sudo gatsby develop -p 80",
         "serve": "gatsby serve",
@@ -99,6 +106,7 @@
         "deploy-branch": "PATH_PREFIX=/deriv-com/br/$branch_name npm run build && gh-pages -d public -e br/$branch_name",
         "prepare": "husky install",
         "pre-commit-lint": "lint-staged",
+        "type-check": "tsc --noEmit",
         "storybook": "start-storybook -p 6006",
         "build:storybook": "build-storybook -c .storybook -o public/storybook"
     },
