+++ tsconfig.json
@@ -1,0 +1,14 @@
+{
+    "include": ["./src/**/*"],
+      "jsx": "react",
+      "compilerOptions": {
+        "baseUrl": "src",
+        "jsx": "preserve",
+        "esModuleInterop": true,
+        "experimentalDecorators": true,
+        "emitDecoratorMetadata": true,
+        "noEmit": true,
+        "skipLibCheck": true,
+        "moduleResolution": "node"
+      }
+}
